pyproject.toml
[project]
name = "alexis"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.12"

package-mode = false

dependencies = [
    "matplotlib (>=3.7.0,<4.0.0)",
    "click (>=8.1.3,<9.0.0)",
    "tqdm (>=4.44.2,<5.0.0)",
    "transformers (>=4.44.2,<5.0.0)",
    "opencv-python (>=4.7.0.72,<5.0.0.0)",
    "attributes-to-language @ git+https://github.com/ruflab/attributes-to-language.git@v0.4.0",
    "auto-sbatch @ git+https://github.com/ruflab/auto-sbatch.git@main",
    "migrate-ckpt @ git+https://github.com/ruflab/migrate-ckpt.git@v0.2.0",
    "cfg-tools @ git+https://github.com/ruflab/cfg-tools.git@main",
    "pillow (>=10.3.0,<11.0.0)",
    "numpy (>=1.26,<2.0)",
    "torch (>=2.0.1,<3.0.0)",
    "wandb (>=0.18.3,<1.0.0)",
    "lightning (>=2.1.0,<3.0.0)",
    "pydantic (>=2.6.0,<3.0.0)",
    "torchvision (>=0.15.2,<1.0.0)",
    "rich (>=13.4.2,<14.0.0)",
    "tokenizers (>=0.19.1,<1.0.0)",
    "ruamel-yaml (>=0.18.6,<1.0.0)",
    "tensorboard (>=2.20.0,<3.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "notebook (>=7.5.4,<8.0.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "shimmer-ssd @ git+https://github.com/AEmanuelli/shimmer-ssd.git",
    "seaborn (>=0.13.2,<0.14.0)",
    "statsmodels (>=0.14.6,<0.15.0)",
    "scikit-posthocs (>=0.12.0,<0.13.0)",
]

[tool.poetry.group.dev.dependencies]
pdoc = "^15.0.1"
pytest = "^7.3.2"
mypy = "1.10.0"
ruff = "^0.6.0"
pre-commit = "^3.6.2"
types-tqdm = "^4.65.0.1"
lxml-stubs = "^0.4.0"
pandas-stubs = "^2.0.2.230605"
types-pillow = "^9.5.0.4"
types-pywin32 = "^306.0.0.2"
types-setuptools = "^67.8.0.0"
types-tabulate = "^0.9.0.2"
types-dataclasses = "^0.6.6"
types-click = "^7.1.8"
types-pyyaml = "^6.0.12.20240311"
types-protobuf = "^4.24.0.20240311"
types-requests = "^2.31.0.20240403"
types-six = "^1.16.21.20240311"
types-redis = "^4.6.0.20240819"
jupyterlab = "^4.3.2"
pyqt5 = "^5.15.11"
ipympl = "^0.9.4"


[tool.ruff]
target-version = "py311"
extend-exclude = [
    "__pycache__",
    ".mypy_cache",
    ".github",
    "data",
]


[tool.ruff.lint]
select = ["E", "W", "F", "UP", "B", "SIM", "I"]

[tool.ruff.lint.isort]
known-third-party = ["wandb", "simple_shapes_dataset"]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
namespace_packages = true
ignore_missing_imports = true
install_types = true
check_untyped_defs = true


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
